+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - Crash when adding a crumb for a timezone change (#3524)
 - Fix a race condition in SentryTracer (#3523)
 - App start ends when first frame is drawn when performanceV2 is enabled (#3530)
+- Use correct rendered frames timestamp for TTID/TTFD and app start (#3531)
+
 - Missing transactions when not calling `reportFullyDisplayed` (#3477)
 
 ## 8.17.2